internal: Move type helpers for schema out to another file
