+++ .changeset/flat-bees-matter.md
@@ -1,0 +1,7 @@
+---
+'@data-client/endpoint': patch
+'@data-client/core': patch
+'@data-client/rest': patch
+---
+
+Change internal organization of some types
