+++ packages/endpoint/tsconfig.compile.json
@@ -1,5 +1,8 @@
 {
   "extends": "./tsconfig",
+  "compilerOptions": {
+    "skipLibCheck": true
+  },
   "include": ["src"],
   "exclude": ["node_modules", "dist", "lib", "legacy", "ts3.4", "**/__tests__"]
 }
